chore: refactor interface helpers
